Remove ecosystem preflight task from pipeline
This is failing currently and I don't want to delay getting the
release-v0.2 build working. IIUC this is not a hard requirement.

Adding a customer EC policy to make this test into only would be a
reasonable idea, but right now we don't have a custom EC policy, so
it seems more pragmatic to just remove it.

EC-464 is filed to bring it back again, (which should be done in
main branch to begin with).

Ref: https://issues.redhat.com/browse/EC-434
Ref: https://issues.redhat.com/browse/EC-464